+++ snuba/datasets/cdc/groupassignee.py
@@ -4,7 +4,7 @@ from snuba.datasets.dataset import Dataset
 from snuba.datasets.plans.single_storage import SingleStorageQueryPlanBuilder
 from snuba.datasets.storages.factory import get_cdc_storage
 from snuba.query.processors.basic_functions import BasicFunctionsProcessor
-from snuba.query.processors.logical import QueryProcessor
+from snuba.query.processors import QueryProcessor
 
 
 class GroupAssigneeDataset(Dataset):
