+++ snuba/datasets/dataset.py
@@ -5,9 +5,9 @@ from snuba.clickhouse.escaping import escape_identifier
 from snuba.datasets.plans.query_plan import StorageQueryPlanBuilder
 from snuba.datasets.storage import Storage, WritableStorage, WritableTableStorage
 from snuba.query.extensions import QueryExtension
-from snuba.query.parsing import ParsingContext
 from snuba.query.logical import Query
-from snuba.query.processors.logical import QueryProcessor
+from snuba.query.parsing import ParsingContext
+from snuba.query.processors import QueryProcessor
 from snuba.util import parse_datetime, qualified_column
