+++ snuba/datasets/outcomes_raw.py
@@ -6,9 +6,9 @@ from snuba.datasets.storages.factory import get_storage
 from snuba.datasets.plans.single_storage import SingleStorageQueryPlanBuilder
 from snuba.query.extensions import QueryExtension
 from snuba.query.organization_extension import OrganizationExtension
+from snuba.query.processors import QueryProcessor
 from snuba.query.processors.basic_functions import BasicFunctionsProcessor
 from snuba.query.processors.timeseries_column_processor import TimeSeriesColumnProcessor
-from snuba.query.processors.logical import QueryProcessor
 from snuba.query.timeseries import TimeSeriesExtension
