+++ snuba/datasets/tags_column_processor.py
@@ -8,15 +8,15 @@ from dataclasses import dataclass
 from snuba import state
 from snuba.clickhouse.columns import ColumnSet
 from snuba.clickhouse.escaping import escape_identifier
-from snuba.query.expressions import Column, Expression, FunctionCall, Literal
 from snuba.query.conditions import (
     BooleanFunctions,
     ConditionFunctions,
     is_binary_condition,
     is_in_condition,
 )
-from snuba.query.parsing import ParsingContext
+from snuba.query.expressions import Column, Expression, FunctionCall, Literal
 from snuba.query.logical import Query
+from snuba.query.parsing import ParsingContext
 from snuba.util import (
     alias_expr,
     escape_literal,
