+++ snuba/query/columns.py
@@ -7,10 +7,10 @@ import _strptime  # NOQA fixes _strptime deferred import issue
 
 from snuba import state
 from snuba.clickhouse.escaping import escape_alias, NEGATE_RE
+from snuba.query.logical import Query
 from snuba.query.parser.conditions import parse_conditions
 from snuba.query.parser.functions import parse_function
 from snuba.query.parsing import ParsingContext
-from snuba.query.logical import Query
 from snuba.query.schema import POSITIVE_OPERATORS
 from snuba.util import (
     alias_expr,
