+++ snuba/query/processors/readonly_events.py
@@ -1,7 +1,7 @@
 from snuba import state
+from snuba.clickhouse.processors import QueryProcessor
 from snuba.clickhouse.query import Query
 from snuba.datasets.schemas.tables import TableSource
-from snuba.clickhouse.processors import QueryProcessor
 from snuba.request.request_settings import RequestSettings
