+++ snuba/query/processors/transaction_column_processor.py
@@ -1,6 +1,6 @@
-from snuba.query.expressions import Column, Expression, FunctionCall, Literal
 from snuba.clickhouse.query import Query
 from snuba.clickhouse.processors import QueryProcessor
+from snuba.query.expressions import Column, Expression, FunctionCall, Literal
 from snuba.request.request_settings import RequestSettings
