+++ snuba/web/db_query.py
@@ -13,9 +13,9 @@ from typing import (
 from sentry_sdk.api import configure_scope
 
 from snuba import settings, state
-from snuba.clickhouse.query import Query
 from snuba.clickhouse.errors import ClickhouseError
 from snuba.clickhouse.formatter import ClickhouseQueryFormatter
+from snuba.clickhouse.query import Query
 from snuba.environment import reader
 from snuba.reader import Result
 from snuba.redis import redis_client
