+++ tests/datasets/test_events.py
@@ -3,8 +3,8 @@ from copy import deepcopy
 from tests.base import BaseEventsTest
 
 from snuba.query.columns import column_expr
-from snuba.query.parsing import ParsingContext
 from snuba.query.logical import Query
+from snuba.query.parsing import ParsingContext
 from snuba.util import tuplify
