+++ tests/query/processors/test_transaction_column_processor.py
@@ -2,10 +2,10 @@ from snuba.clickhouse.columns import ColumnSet
 from snuba.clickhouse.formatter import ClickhouseExpressionFormatter
 from snuba.datasets.schemas.tables import TableSource
 from snuba.query.expressions import Column, FunctionCall, Literal
+from snuba.query.logical import Query
 from snuba.query.processors.transaction_column_processor import (
     TransactionColumnProcessor,
 )
-from snuba.query.logical import Query
 from snuba.request.request_settings import HTTPRequestSettings
