+++ tests/query/test_organization_extension.py
@@ -5,8 +5,8 @@ from snuba.clickhouse.columns import ColumnSet
 from snuba.datasets.schemas.tables import TableSource
 from snuba.query.conditions import binary_condition, ConditionFunctions
 from snuba.query.expressions import Column, Literal
-from snuba.query.organization_extension import OrganizationExtension
 from snuba.query.logical import Query
+from snuba.query.organization_extension import OrganizationExtension
 from snuba.request.request_settings import HTTPRequestSettings
 from snuba.schemas import validate_jsonschema
