+++ tests/query/test_timeseries_extension.py
@@ -11,8 +11,8 @@ from snuba.query.conditions import (
     ConditionFunctions,
 )
 from snuba.query.expressions import Column, Expression, Literal
-from snuba.query.timeseries import TimeSeriesExtension
 from snuba.query.logical import Query
+from snuba.query.timeseries import TimeSeriesExtension
 from snuba.query.types import Condition
 from snuba.request.request_settings import HTTPRequestSettings
 from snuba.schemas import validate_jsonschema
