+++ tests/test_util.py
@@ -12,8 +12,8 @@ from snuba.query.columns import (
     complex_column_expr,
     conditions_expr,
 )
-from snuba.query.parsing import ParsingContext
 from snuba.query.logical import Query
+from snuba.query.parsing import ParsingContext
 from snuba.util import (
     escape_literal,
     tuplify,
